x86/msr: Definitions for MSR_INTEL_CORE_THREAD_COUNT
This is a model specific register which details the current configuration
cores and threads in the package.  Because of how Hyperthread and Core
configuration works works in firmware, the MSR it is de-facto constant and
will remain unchanged until the next system reset.

It is a read only MSR, and for now, reject guest attempts to read it, to avoid
the system setting leaking into guest context.  Further CPUID/MSR work is
required before we can start virtualising a consistent topology to the guest.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Message-Id: <1555172532-23814-3-git-send-email-andrew.cooper3@citrix.com>
Acked-by: Jan Beulich <jbeulich@suse.com>