xen/arm: page: Clarify the Xen TLBs helpers name
Now that we dropped flush_xen_text_tlb_local(), we have only one set of
helpers acting on Xen TLBs. There naming are quite confusing because the
TLB instructions used will act on both Data and Instruction TLBs.

Take the opportunity to rework the documentation that can be confusing
to read as they don't match the implementation.

Lastly, switch from unsigned lont to vaddr_t as the function technically
deal with virtual address.

Signed-off-by: Julien Grall <julien.grall@arm.com>
Message-Id: <20190417175815.16905-5-julien.grall@arm.com>
Reviewed-by: Andrii Anisov <andrii_anisov@epam.com>