xen/arm: Allow cleaning the directory even when CONFIG_EARLY_PRINTK is set
CONFIG_EARLY_PRINTK can only be set when CONFIG_DEBUG is enabled. It can
be quite convenient to only modify the target.

However, the target clean will not include the .config. This means
CONFIG_DEBUG is not enabled and therefore make will throw an error
preventing clean to continue.

The check is not moved at linking time.

Signed-off-by: Julien Grall <julien.grall@arm.com>

Message-Id: <20190422164106.20331-1-julien.grall@arm.com>