x86/pt: skip setup of posted format IRTE when gvec is 0
When testing with an UP guest with a pass-thru device with vt-d pi
enabled in host, we observed that guest couldn't receive interrupts
from that pass-thru device. Dumping IRTE, we found the corresponding
IRTE is set to posted format with "vector" field as 0.

We would fall into this issue when guest used the pirq format of MSI
(see the comment xen_msi_compose_msg() in linux kernel). As 'dest_id'
is repurposed, skip migration which is based on 'dest_id'.

Signed-off-by: Chao Gao <chao.gao@intel.com>
Message-Id: <1556601559-30921-1-git-send-email-chao.gao@intel.com>