xen/arm: keep track of reserved-memory regions
As we parse the device tree in Xen, keep track of the reserved-memory
regions as they need special treatment (follow-up patches will make use
of the stored information.)

Reuse process_memory_node to add reserved-memory regions to the
bootinfo.reserved_mem array. Remove the warning if there is no reg in
process_memory_node because it is a normal condition for
reserved-memory.

Signed-off-by: Stefano Stabellini <stefanos@xilinx.com>

Message-Id: <1556658172-8824-8-git-send-email-sstabellini@kernel.org>