xen/sched: Change vcpu_migrate_*() to operate on schedule item
Now that vcpu_migrate_start() and vcpu_migrate_finish() are used only
to ensure a vcpu is running on a suitable processor they can be
switched to operate on schedule items instead of vcpus.

While doing that rename them accordingly and make the _start() variant
static.

vcpu_move_locked() is switched to schedule item, too.

Signed-off-by: Juergen Gross <jgross@suse.com>
Message-Id: <20190506065644.7415-32-jgross@suse.com>