xen/sched: add fall back to idle vcpu when scheduling item
When scheduling an item with multiple vcpus there is no guarantee all
vcpus are available (e.g. above maxvcpus or vcpu offline). Fall back to
idle vcpu of the current cpu in that case. This requires to store the
correct schedule_item pointer in the idle vcpu as long as it used as
fallback vcpu.

In order to modify the runstates of the correct vcpus when switching
schedule items merge sched_item_runstate_change() into
sched_switch_items() and loop over the affected physical cpus instead
of the item's vcpus. This in turn requires an access function to the
current variable of other cpus.

Today context_saved() is called in case previous and next vcpus differ
when doing a context switch. With an idle vcpu being capable to be a
substitute for an offline vcpu this is problematic when switching to
an idle scheduling item. An idle previous vcpu leaves us in doubt which
schedule item was active previously, so save the previous item pointer
in the per-schedule resource area and use its value being non-NULL as
a hint whether context_saved() should be called.

Signed-off-by: Juergen Gross <jgross@suse.com>
Message-Id: <20190506065644.7415-43-jgross@suse.com>