xen/watchdog: Rearrange the logic to fold the timer-arming paths
By rearranging the logic, the timer allocation loop can reuse the common timer
arming/clearing logic.  This results in easier to follow code, and a modest
reduction in compiled code size (-64, 290 => 226).

For domains which use watchdogs, the overwhemling majoriy of hypercalls will
be re-arming an existing timer.  Arrange the fastpath to match.

This does cause one change in behaviour for a corner case.  Previously,
specifying id = 0, timeout = 0 would instantly kill the domain, as the timer
would fire before returning to the guest.  This corner case is going to be
reused for a different purpose in a later change.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Message-Id: <1557512884-32395-3-git-send-email-andrew.cooper3@citrix.com>