xen/watchdog: Drop all locked operations on the watchdog_inuse_map
All modifications to the watchdog_inuse_map happen with d->watchdog_lock held,
so there are no concurrency problems to deal with.

Furthermore, there is no need to use a loop to locate the next available
watchdog.  As the bitmap is currently 2 bits wide and is stored in a uint32_t,
the next available timer can be located in O(1) time using bit-scanning
instructions.

No change in behaviour, but should have less cache-coherency impact.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Message-Id: <1557512884-32395-4-git-send-email-andrew.cooper3@citrix.com>