libxc: elf_kernel loader: Remove check for shstrtab
This was probably useful to load ELF Note, but now ELF notes
"should live in a PT_NOTE segment" (elfnote.h).

With notes living in segment, there are no need for sections, so there
is nothing to be stored in the shstrtab.

This patch would allow to write a simpler ELF header for an OVMF blob
(which isn't an ELF) and allow it to be loaded as a PVH kernel. The
header only needs to declare two program segments:
- one to tell an ELF loader where to put the blob,
- one for a Xen ELFNOTE.

The ELFNOTE is to comply to the pvh design which wants the
XEN_ELFNOTE_PHYS32_ENTRY to declare a blob as compaptible with the PVH
boot ABI.

Note that without the ELFNOTE, libxc will load an ELF but with
the plain ELF loader, which doesn't check for shstrtab.

Signed-off-by: Anthony PERARD <anthony.perard@citrix.com>
Message-Id: <20190515114015.25492-1-anthony.perard@citrix.com>
Acked-by: Wei Liu <wei.liu2@citrix.com>