hotplug/Linux: fix starting of xenstored with systemd
A hard to trigger race with another unrelated systemd service and
xenstored.service unveiled a bug in the way how xenstored is launched
with systemd.

launch-xenstore may start either a daemon or a domain. In case a domain
is used, systemd-notify was called. If another service triggered a
restart of systemd while xenstored.service was executed, systemd may
temporary lose track of services with Type=notify. As a result,
xenstored.service would be marked as failed and units that depend on it
will not be started anymore. This breaks the enire Xen toolstack.

Currently the decision which variant of xenstore should be used is
controlled via /etc/sysconfig/xencommons:XENSTORETYPE=[domain|daemon].
This change preserves this functionality for the sysv and systemd.

One way to fix it is to handle the domain case as Type=oneshot because
there is nothing to monitor for systemd. The daemon case has to be
handled as Type=simple, which is the default if no Type= is specified.
A single unit can have just one type, so a new unit for the daemon case
has to be created to preserve existing setups during upgrading of Xen.
This new xenstored-daemon.service is started on demand by the existing
xenstored.service. Since it is a separate unit, systemd will supervise
the xenstored daemon in dom0.

launch-xenstore expects now two arguments, the type of the init
system, and the optional xenstore type. The systemd-notify calls are
removed because systemd does not expect any notification with the
updated .service files. In case the init system is systemd the daemon
or init-xenstore-domain is started via exec to make sure systemd
monitors the process it has just launched. Various things specific to
either sysv or systemd are now handled separately.

The start of xenstored-daemon.service in launch-xenstore will
essentially call this helper twice. A separate internal state is
introduced to deal with this. This is required to handle the xencommons
case explained above.

A followup change will remove the code which calls to sd_notify, they
are not needed after the separation of Type=oneshot for domain and
Type=simple for daemon.

Signed-off-by: Olaf Hering <olaf@aepfle.de>
Message-Id: <20190515150632.16269-1-olaf@aepfle.de>