coverage: GCC coverage libfdt Makefile fix
The patch resolves 'xencov' crashes in case of Aarch64.

All the .init.* sections are stripped after boot,
it means that anything in .init.data cannot be accessed anymore.
The build system explicitly compiles any .init binary without gcov option.
The problem is coming from libfdt.
The entire library is moved to .init using:
$(OBJCOPY) $(foreach s,$(SECTIONS),--rename-section .$(s)=.init.$(s)) $< $@
So we need to tell the top Makefile to filter out libfdt.

Reported-by: Viktor Mitin <viktor.mitin.19@gmail.com>
Signed-off-by: Julien Grall <julien.grall@arm.com>
Tested-by: Viktor Mitin <viktor.mitin.19@gmail.com>
Message-Id: <20190516111354.15195-1-viktor.mitin.19@gmail.com>
Acked-by: Andrew Cooper <andrew.cooper3@citrix.com>
Reviewed-by: Wei Liu <wei.liu2@citrix.com>