hotplug/Linux: fix starting of xenstored with restarting systemd
A hard to trigger race with another unrelated systemd service and
xenstored.service unveiled a bug in the way how xenstored is launched
with systemd.

launch-xenstore may start either a daemon or a domain. In case a domain
is used, systemd-notify was called. If another service triggered a
restart of systemd while xenstored.service was executed, systemd may
temporary lose track of services with Type=notify. As a result,
xenstored.service would be marked as failed and units that depend on it
will not be started anymore. This breaks the enire Xen toolstack.

The chain of events is basically: xenstored.service sends the
notification to systemd, this is a one-way event. Then systemd may be
restarted by the other unit. During this time xenstored.service is done
and exits. Once systemd is done with its restart it collects the pending
notifications and childs. If it does not find the unit which sent the
notification it will declare it as failed.

A workaround for this scenario is to wait for a short time after sending
to notification. If systemd happens to restart it will still find the
unit it launched.

Adjust the callers of launch-xenstore to specifiy the init system.
Do not fork xenstored with systemd, preserve pid.
Be verbose about xenstored startup only with sysv to avoid interleaved
output in systemd journal. Do the same also for domain case, even if is
not strictly needed because init-xenstore-domain has no output.

The fix for upstream systemd which is supposed to fix it:
575b300b795b6 ("pid1: rework how we dispatch SIGCHLD and other signals")

v02:
- preserve Type=notify

Signed-off-by: Olaf Hering <olaf@aepfle.de>
Message-Id: <20190517095621.24271-1-olaf@aepfle.de>