xen/mm: Convert {s, g}et_gpfn_from_mfn() to use typesafe MFN
The first parameter of {s,g}et_gpfn_from_mfn() is an MFN, so it can be
switched to use the typesafe.

At the same time, replace gpfn with pfn in the helpers as they all deal
with PFN and also turn the macros to static inline.

Note that the return of the getter and the 2nd parameter of the setter
have not been converted to use typesafe PFN because it was requiring
more changes than expected.

Signed-off-by: Julien Grall <julien.grall@arm.com>
Acked-by: Tamas K Lengyel <tamas@tklengyel.com>
Acked-by: Stefano Stabellini <sstabellini@kernel.org>

Message-Id: <20190603160350.29806-14-julien.grall@arm.com>