xen/link: Fold .data.read_mostly into .data
.data.read_mostly only needs separating from adjacent data by one cache line
to be effective, and placing it adjacent to .data.page_aligned fulfills this
requirement.

For ARM, .ex_table appears to be a vestigial remnant.  Nothing in the
resulting build ever inspects or acts on the contents of the table.  The arm32
build does however have some assembly routines which fill .ex_table.

Drop all of ARM's .ex_table infrastructure, to reduce the size of the compiled
binary, and avoid giving the illusion of exception handling working.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Message-Id: <1560975087-25632-4-git-send-email-andrew.cooper3@citrix.com>