x86: Introduce Xen detection as separate logic from Xen Guest support.
Add Kconfig option XEN_DETECT for:
  "Support for Xen detecting when it is running under Xen".
If running under Xen is detected, a boot message will indicate the
hypervisor version obtained from cpuid.

Update the XEN_GUEST Kconfig option text to reflect its current
purpose:
  "Common PVH_GUEST and PV_SHIM logic for Xen as a Xen-aware guest".

Update calibrate_APIC_clock to use Xen-specific init if nested Xen is
detected, even if not operating as a PV shim or booted as PVH.

This work is a precursor to adding the interface for support of
PV drivers on nested Xen.

Signed-off-by: Christopher Clark <christopher.clark@starlab.io>
Message-Id: <20190620003053.21993-3-christopher.w.clark@gmail.com>