x86/nested: add nested_xen_version hypercall
Provides proxying to the host hypervisor for XENVER_version and
XENVER_get_features ops.

The nested PV interface is only enabled when Xen is not running as
either the PV shim or booted as PVH, since the initialization performed
within the hypervisor in those cases - ie. as a Xen guest - claims resources
that are normally operated by the control domain.

This nested hypercall only permits access from the control domain.
The XSM policy hook implementation is deferred to a subsequent commit.

Signed-off-by: Christopher Clark <christopher.clark@starlab.io>
Message-Id: <20190620003053.21993-4-christopher.w.clark@gmail.com>