x86/nested, xsm: add nested_event_channel_op hypercall
Provides proxying to the host hypervisor for these event channel ops:
 * EVTCHNOP_alloc_unbound
 * EVTCHNOP_bind_vcpu
 * EVTCHNOP_close
 * EVTCHNOP_send
 * EVTCHNOP_unmask

Introduces a new XSM access vector class for policy control applied to this
operation: nested_event.
This is required because the existing 'event' access vector is unsuitable
for repurposing to the nested case: it operates on per-channel security
identifiers that are generated from a combination of the security
identifiers of the two communicating endpoints and data is not available for
the remote endpoint in the nested case.

Signed-off-by: Christopher Clark <christopher.clark@starlab.io>
Message-Id: <20190620003053.21993-9-christopher.w.clark@gmail.com>